refactor(mcp): validate search_transactions label ids and surface labels in the tool description
Reuse labelsInSpace (promoted to the McpTool base so read and write tools
share it) so an unknown or stale label id returns the same actionable
'call list_labels' error the rest of the MCP surface gives, instead of a
silent empty result. Mention label filtering in the tool description so
agents discover the capability.
