+++ app/Mcp/Tools/McpTool.php
@@ -3,8 +3,10 @@
 namespace App\Mcp\Tools;
 
 use App\Enums\PlanFeature;
+use App\Models\Label;
 use App\Models\Space;
 use App\Models\User;
+use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Support\Str;
 use Illuminate\Validation\ValidationException;
 use Laravel\Mcp\Request;
@@ -100,5 +102,37 @@ abstract class McpTool extends Tool
         }
 
         return $space;
+    }
+
+    /**
+     * Resolve every label id passed under $key, asserting each belongs to the
+     * space. Returns an empty collection when the argument is absent or empty.
+     *
+     * @return Collection<int, Label>
+     */
+    protected function labelsInSpace(Request $request, Space $space, string $key): Collection
+    {
+        $ids = collect($request->get($key, []))
+            ->map(fn (mixed $id): string => (string) $id)
+            ->filter()
+            ->unique()
+            ->values();
+
+        if ($ids->isEmpty()) {
+            /** @var Collection<int, Label> $empty */
+            $empty = Label::query()->whereRaw('1 = 0')->get();
+
+            return $empty;
+        }
+
+        $labels = Label::query()->forSpace($space)->whereIn('id', $ids)->get();
+
+        if ($labels->count() !== $ids->count()) {
+            throw ValidationException::withMessages([
+                $key => "One or more label ids do not exist in space {$space->id}. Call list_labels to see valid ids.",
+            ]);
+        }
+
+        return $labels;
     }
 }
